Fix /scan 500 on runtimes without zbar; log unhandled errors
- cv2.QRCodeDetector fallback returned float corner points, which
  cv2.line rejects (cv2.error) when annotate_image drew bounding boxes
  after an upload -> 500. Return integer points instead.
- Harden annotate_image against drawing errors.
- Add a global Flask error handler that logs the traceback to the
  gunicorn logger (visible in Render logs) and returns JSON errors.

+++ core/detector.py
@@ -85,7 +85,7 @@ class QRDetector:
                         text = text.decode("utf-8", errors="replace")
                     pts = []
                     if points is not None and len(points) > i:
-                        pts = [(float(x), float(y)) for x, y in points[i]]
+                        pts = [(int(x), int(y)) for x, y in points[i]]
                     results.append(DetectedCode(data=text, points=pts))
         except Exception:
             pass
